+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Experimental fuzzy matching for package namespace completions (`experimental.fuzzy_namespace_completion`)
+- Experimental fuzzy matching for package namespace completions (`experimental.completion_namespace.fuzzy`)
 
 ## [0.2.3] - 2026-02-27